Module Yocaml_unix

Unix runtime for YOCaml.

Allows you to run YOCaml on a Unix system based on Unix (or source service for more complex runtimes).

val run : ?level:Yocaml_runtime.Log.level -> ?custom_error_handler: (Format.formatter -> Yocaml.Data.Validation.custom_error -> unit) -> (unit -> unit Yocaml.Eff.t) -> unit

run ?custom_error_handler program Runs a Yocaml program in the Unix runtime. The log level (default: Debug) and a custom_error_handler can be passed as arguments to change the reporting level and use a default log reporter, and to pretty print user defined errors. If the user want to set-up a custom reporter, the level flag should be ignored.

val serve : ?level:Yocaml_runtime.Log.level -> ?custom_error_handler: (Format.formatter -> Yocaml.Data.Validation.custom_error -> unit) -> target:Yocaml.Path.t -> port:int -> (unit -> unit Yocaml.Eff.t) -> unit

serve ?level ?custom_error_handler ~target ~port program serve the directory target statically and re-run program on each refresh. If the user want to set-up a custom reporter, the level flag should be ignored.

Runtime

module Runtime : sig ... end

Rexports Runtime to be used outside of the package.