Module Yocaml_unix__.Runtime

include Yocaml.Runtime.RUNTIME

val get_time : unit -> float

get_time () should returns a float like Unix.gettimeofday ().

val file_exists : Yocaml.Filepath.t -> bool

file_exists path should returns true if path exists (as a file or a directory), false otherwise.

val target_exists : Yocaml.Filepath.t -> bool

Same of file_exists but acting on the target.

val is_directory : Yocaml.Filepath.t -> bool

is_directory path should returns true if path is an existing file and if the file is a directory, false otherwise.

val get_modification_time : Yocaml.Filepath.t -> int Yocaml.Try.t

get_modification_time path should returns a Try.t containing the modification time (as an integer) of the given file. The function may fail.

val target_modification_time : Yocaml.Filepath.t -> int Yocaml.Try.t

Same of get_modification_time but acting on the target.

val read_file : Yocaml.Filepath.t -> string Yocaml.Try.t

read_file path should returns a Try.t containing the content (as a string) of the given file. The function may fail.

val content_changes : Yocaml.Filepath.t -> string -> bool Yocaml.Try.t

content_changes filepath new_content check if the content of the file has been changed.

val write_file : Yocaml.Filepath.t -> string -> unit Yocaml.Try.t

write_file path content should write (create or overwrite) content into the given path. The function may fail.

val read_dir : Yocaml.Filepath.t -> Yocaml.Filepath.t list

read_dir path should returns a list of children. The function is pretty optimistic if the directory does not exist, or for any other possible reason the function should fail, it will return an empty list.

val create_dir : ?⁠file_perm:int -> Yocaml.Filepath.t -> unit

create_dir path is an optimistic version of mkdir -p, the function extract the directory of a file and create it if it does not exists without any failure.

val log : Yocaml.Log.level -> string -> unit

log level message justs dump a message on stdout.