Module Yocaml_unix__.Runtime

include Yocaml.Runtime.RUNTIME

val file_exists : Yocaml.Aliases.filepath -> bool

file_exists path should returns true if path exists (as a file or a directory), false otherwise.

val is_directory : Yocaml.Aliases.filepath -> bool

is_directory path should returns true if path is an existing file and if the file is a directory, false otherwise.

val get_modification_time : Yocaml.Aliases.filepath -> int Yocaml.Try.t

get_modification_time path should returns a Try.t containing the modification time (as an integer) of the given file. The function may fail.

val read_file : Yocaml.Aliases.filepath -> string Yocaml.Try.t

read_file path should returns a Try.t containing the content (as a string) of the given file. The function may fail.

val write_file : Yocaml.Aliases.filepath -> string -> unit Yocaml.Try.t

write_file path content should write (create or overwrite) content into the given path. The function may fail.

val read_dir : Yocaml.Aliases.filepath -> Yocaml.Aliases.filepath list

read_dir path should returns a list of children. The function is pretty optimistic if the directory does not exist, or for any other possible reason the function should fail, it will return an empty list.

val create_dir : ?⁠file_perm:int -> Yocaml.Aliases.filepath -> unit

create_dir path is an optimistic version of mkdir -p, the function extract the directory of a file and create it if it does not exists without any failure.

val log : Yocaml.Aliases.log_level -> string -> unit

log level message justs dump a message on stdout.